requirements.txt
Django>=5.0
django-ninja>=1.1
python-dotenv>=1.0
lxml>=5.0
geopandas>=0.14
shapely>=2.0
pyproj>=3.6
psycopg2-binary>=2.9